feat: Accept command-line arguments
Now we can pass files through command-line arguments.

**Other changes:**
- Add .clang-format for nice format
- Add .gitignore
- Add CMake configs
- Split declarations from source files
- Organize project structure
