+++ compile_commands.json
@@ -1,0 +1,1 @@
+build/compile_commands.json
