+++ example/source.tbs
@@ -1,0 +1,3 @@
+let x = 5;
+
+for (let i = 0; i < 5; i++) {};
