--- source.tbs
@@ -1,2 +1,0 @@
-let x = 5;
- for(let i = 0; i < 5; i++) {};
