+++ src/CMakeLists.txt
@@ -1,0 +1,1 @@
+target_sources(${EXECUTABLE_NAME} PRIVATE main.cpp utils.cpp lexer.cpp)
