+++ src/utils.h
@@ -1,0 +1,9 @@
+#ifndef SRC_UTILS_H_
+#define SRC_UTILS_H_
+
+#include <string>
+#include <string_view>
+
+std::string readFile(std::string_view file_name);
+
+#endif  // SRC_UTILS_H_
